move types to database
